% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_1_Generate_reports.R
\docType{data}
\name{sites}
\alias{sites}
\title{Sites data of Danube dike plots}
\format{
A data frame with 1152 rows and 16 variables:
}
\usage{
sites
}
\description{
A dataset containing the sites information
}
\keyword{datasets}
